check APIStatus.Code in Is* family of functions
Check the Code in the Is family of functions where it makes sense. This
was already done in a couple of functions, but there were many others
where checking the code makes sense.

IsConflict has to check IsAlreadyExists, because IsAlreadyExists uses
http.StatusConflict as the error.

IsTimeout has to check IsServerTimeout, because IsServerTimeout also
uses http.StatusGatewayTimeout.